Make pkg-deb use the source package generated by pkg-src.

+++ tool/pkg.py
@@ -6,7 +6,7 @@
 
 from job import job, log, debug, fatal, warn, run, rmtree, mktree
 from vm import LinuxBenchVM, WindowsBenchVM, DATA_ROOT, CTL_DIR
-import os, glob
+import os, glob, re
 
 
 src_vm = LinuxBenchVM('src', 'debian', 22)
@@ -59,23 +59,36 @@ def pkg_deb():
     This job creates the debian source and binary packages.
     """
     if deb_vm.missing():
-       raise fatal("VM deb is not built")
+        raise fatal("VM is not built: %s" % deb_vm.name)
+    source_paths = glob.glob("./build/pkg/src/HTSQL-*.tar.gz")
+    if not source_paths:
+        raise fatal("cannot find a source package; run `job pkg-src` first")
+    if len(source_paths) > 1:
+        raise fatal("too many source packages in `./build/pkg/src`")
+    [source_path] = source_paths
+    archive = os.path.basename(source_path)
+    version = re.match(r'^HTSQL-(?P<version>.+).tar.gz$',
+                       archive).group('version')
     if deb_vm.running():
-       deb_vm.stop()
+        deb_vm.stop()
     if os.path.exists("build/pkg/deb"):
-       rmtree("build/pkg/deb")
+        rmtree("build/pkg/deb")
     deb_vm.start()
-    src_path = deb_vm.download([PYPISRC])
-    deb_vm.put(src_path, '/root')
-    deb_vm.run("mv /root/%s /root/htsql_%s.orig.tar.gz" % (ARCHIVE, VERSION))
-    deb_vm.run("tar xvfz /root/htsql_%s.orig.tar.gz -C /root" % VERSION)
-    deb_vm.put(DATA_ROOT+"/pkg/debian", "/root/HTSQL-%s" % VERSION)
-    deb_vm.run("cd /root/HTSQL-%s && dpkg-buildpackage" % VERSION)
-    mktree("build/pkg/deb")
-    deb_vm.get("/root/htsql_%s-1_all.deb" % VERSION, "build/pkg/deb")
-    deb_vm.run("dpkg -i /root/htsql_%s-1_all.deb" % VERSION)
-    deb_vm.run("htsql-ctl get sqlite:test.db /%27Hello%20World%27")
-    deb_vm.stop()
+    try:
+        deb_vm.put(source_path, '/root')
+        deb_vm.run("mv %s htsql_%s.orig.tar.gz" % (archive, version))
+        deb_vm.run("tar xvfz htsql_%s.orig.tar.gz" % version)
+        deb_vm.put(DATA_ROOT+"/pkg/debian", "HTSQL-%s" % version)
+        deb_vm.run("cd HTSQL-%s && dpkg-buildpackage" % version)
+        mktree("build/pkg/deb")
+        deb_vm.get("htsql_%s-1_all.deb" % version, "build/pkg/deb")
+        deb_vm.run("dpkg -i htsql_%s-1_all.deb" % version)
+        deb_vm.run("htsql-ctl get sqlite:test.db /%27Hello%20World%27")
+    finally:
+        deb_vm.stop()
     log()
-    log("`build/pkg/deb/htsql_%s-1_all.deb` is created successfully" % VERSION)
+    log("Debian package is built successfully:")
+    log("  `./build/pkg/deb/htsql_%s-1_all.deb`" % version)
+    log()
+
